# specs/001-heliolus-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Heliolus Compliance Assessment Platform API
  description: API for compliance assessment, gap analysis, and vendor marketplace
  version: 1.0.0
  contact:
    name: Heliolus Support
    email: support@heliolus.com
servers:
  - url: https://api.heliolus.com/v1
    description: Production
  - url: https://staging-api.heliolus.com/v1
    description: Staging
  - url: http://localhost:3001/v1
    description: Development
security:
  - bearerAuth: []
...
paths:
...
  /documents/{id}:
    get:
      tags: [Documents]
      summary: Get document
      operationId: getDocument
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        404:
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Documents]
      summary: Delete document
      operationId: deleteDocument
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        204:
          description: Document deleted
        404:
          $ref: '#/components/responses/NotFound'

  # Templates
...
  /assessments:
    get:
      tags: [Assessments]
      summary: List assessments
      operationId: listAssessments
      parameters:
        - name: organizationId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AssessmentStatus'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        200:
          description: Assessment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Assessment'

    post:
      tags: [Assessments]
      summary: Create assessment
      operationId: createAssessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [organizationId, templateId]
              properties:
                organizationId:
                  type: string
                templateId:
                  type: string
      responses:
        201:
          description: Assessment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
        400:
          $ref: '#/components/responses/BadRequest'
        402:
          description: Insufficient credits
  /assessments/{id}:
    get:
      tags: [Assessments]
      summary: Get assessment
      operationId: getAssessment
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Assessment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
        404:
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Assessments]
      summary: Update assessment
      operationId: updateAssessment
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                responses:
                  type: object
                status:
                  $ref: '#/components/schemas/AssessmentStatus'
      responses:
        200:
          description: Assessment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
...
    Document:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        documentType:
          $ref: '#/components/schemas/DocumentType'
        createdAt:
          type: string
          format: date-time
    DocumentType:
      type: string
      enum: [POLICY, ANNUAL_REPORT, COMPLIANCE_CERT, AUDIT_REPORT, OTHER]
...
    Assessment:
      type: object
      properties:
        id:
          type: string
        organizationId:
          type: string
        templateId:
          type: string
        status:
          $ref: '#/components/schemas/AssessmentStatus'
        responses:
          type: object
        riskScore:
          type: number
        creditsUsed:
          type: integer
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    AssessmentStatus:
      type: string
      enum: [DRAFT, IN_PROGRESS, COMPLETED, FAILED]
...